+++ projects/test-app/src/app/dynamic-component/dynamic-component.component.ts
@@ -1,4 +1,4 @@
-import { Component, EventEmitter, OnInit, Output } from '@angular/core';
+import { Component, EventEmitter, Input, OnInit, Output } from '@angular/core';
 
 @Component({
   selector: 'app-dynamic-component',
@@ -8,6 +8,9 @@ import { Component, EventEmitter, OnInit, Output } from '@angular/core';
 export class DynamicComponentComponent implements OnInit {
 
   @Output() destroyToast: EventEmitter<boolean> = new EventEmitter<boolean>(false);
+  @Input() context!: {
+    content: string
+  };
 
   constructor() { }
 
@@ -18,6 +21,8 @@ export class DynamicComponentComponent implements OnInit {
   }
 
   ngOnInit(): void {
+    console.log('Dynamically injected data:')
+    console.log(this.context.content);
   }
 
 }